requirements.txt
streamlit
numpy
sentence-transformers
faiss-cpu 